+++ eventkit_cloud/jobs/signals.py
@@ -18,11 +18,13 @@ from eventkit_cloud.jobs.models import (
     Region,
     RegionalPolicy,
 )
-from eventkit_cloud.tasks.export_tasks import make_dirs
+from eventkit_cloud.utils.helpers import make_dirs
 from eventkit_cloud.tasks.helpers import make_file_downloadable
-from eventkit_cloud.utils.helpers import clear_mapproxy_config_cache
 from eventkit_cloud.utils.image_snapshot import save_thumbnail
-from eventkit_cloud.utils.mapproxy import get_mapproxy_config_template
+from eventkit_cloud.utils.mapproxy import (
+    get_mapproxy_config_template,
+    clear_mapproxy_config_cache,
+)
 from eventkit_cloud.utils.s3 import delete_from_s3
 
 logger = logging.getLogger(__name__)
