+++ eventkit_cloud/tasks/helpers.py
@@ -1,3 +1,4 @@
+from contextlib import contextmanager
 import copy
 import glob
 import json
@@ -40,8 +41,8 @@ from eventkit_cloud.tasks.enumerations import Directory, PREVIEW_TAIL, UNSUPPORT
 from eventkit_cloud.tasks.exceptions import FailedException
 from eventkit_cloud.tasks.models import DataProviderTaskRecord, ExportRunFile, ExportTaskRecord, ExportRun
 from eventkit_cloud.utils import s3
+from eventkit_cloud.utils.helpers import make_dirs
 from eventkit_cloud.utils.generic import retry
-from eventkit_cloud.utils.helpers import cd, get_file_paths  # NOQA
 from eventkit_cloud.utils.mapproxy import get_chunked_bbox
 from eventkit_cloud.utils.s3 import download_folder_from_s3
 
@@ -1317,9 +1318,24 @@ def make_file_downloadable(file_path: Path, skip_copy: bool = False) -> Tuple[Pa
     return file_name, download_url
 
 
-def make_dirs(path):
+@contextmanager
+def cd(newdir):
+    prevdir = os.getcwd()
+    os.chdir(newdir)
     try:
-        os.makedirs(path, 0o751, exist_ok=True)
-    except OSError:
-        if not os.path.isdir(path):
-            raise
+        yield
+    finally:
+        os.chdir(prevdir)
+
+
+def get_file_paths(directory):
+    """
+    Gets file paths with absolute file paths for copying the files and a relative file path for
+    where the file should be located in the datapack relative to the directory.
+    """
+    paths = {}
+    with cd(directory):
+        for dirpath, _, filenames in os.walk("./"):
+            for f in filenames:
+                paths[os.path.abspath(os.path.join(dirpath, f))] = os.path.join(dirpath, f)
+    return paths
